+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* **dquery** — Added support for `add(c)`, `classes()`, `has(c)`, `remove(c)`, `replace(c1, c2)`.
+
 ## 1.0.0-rc13 - 2022-03-18
 
 * Initial implementation of VDOM.